# REF: split out sorted_rank algo