@jreback 2021-06-10
great. assume perf is the same.