@mzeitlin11 2021-06-10
> great. assume perf is the same.

Yeah perf is posted above in a details block